Revamped AutoMapper with custom profil Initialization and  added AutoMapper Configurator class to utilize the profiles. Please refer to https://github.com/AutoMapper/AutoMapper/wiki/Configuration for more info.
